==> 02_prepare_data_for_modeling.Rmd <==
---
title: "Prepare Data for Modeling"
date: "1/23/2019"
output: html_document
---

```{r}
library(tidyverse)
library(caret)
```


This file serves as a second pass at data processing with the aim to prepare data for
subsequent model fitting. The two primary steps carried out here is to partition the
data into training and testing sets and to scale them to allow for coefficient comparison.

### Load

```{r}
setwd("data/processed/")
historical_data <- read_csv("historical_data.csv",
                            col_types = list(wbic = col_character(),
                                             year = col_character(),
                                             month  = col_character(),
                                             season = col_character(),
                                             summerkill = col_character(),
                                             ice_duration = col_double())) %>%
  select(-cause.category.4, -anthropogenic, -infectious, -unknown, -winterkill) %>%
  mutate(summerkill = fct_recode(summerkill,
                                 "neg" = "0",
                                 "pos" = "1"))

future_data <- read_csv("future_data.csv",
                        col_types = list(wbic = col_character(),
                                         year = col_character(),
                                         month  = col_character(),
                                         season = col_character(),
                                         ice_duration = col_double()))
```

## Partition

```{r}
set.seed(998)
in_training <- createDataPartition(historical_data$summerkill,
                                  p = .75, list = FALSE)
training <- historical_data  %>%
  slice(in_training)
testing  <- historical_data %>%
  slice(-in_training)
```

## Scale

```{r}
pre_proc_values <- preProcess(training, method = c("center", "scale"))
training <- predict(pre_proc_values, training) %>%
  mutate(year = as.integer(year))
testing  <- predict(pre_proc_values, testing) %>%
  mutate(summerkill = factor(testing$summerkill, 
                             levels = c("pos", "neg"), 
                             ordered = TRUE),
         year = as.integer(year))

pre_proc_values_full_hist <- preProcess(historical_data, method = c("center", "scale"))
future  <- predict(pre_proc_values_full_hist, future_data) %>%
  mutate(year = as.integer(year))
```

## Write out data

```{r}
write_csv(training, "../prepared_for_modeling/training.csv")
write_csv(testing, "../prepared_for_modeling/testing.csv")
write_csv(future, "../prepared_for_modeling/future.csv")
```
